pr-sug: move integration types to respective categories
